+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ProductService } from '../../services/product.service';
 import { ProductModel } from '../../models/product.model';
+import { CartService } from 'src/app/cart/services/cart.service';
 
 @Component({
   selector: 'app-product-list',
@@ -10,9 +11,15 @@ import { ProductModel } from '../../models/product.model';
 export class ProductListComponent implements OnInit {
   products!: Array<ProductModel>;
 
-  constructor(private productService: ProductService) { }
+  constructor(
+    private productService: ProductService,
+    private cartService: CartService) { }
 
   ngOnInit(): void {
     this.products = this.productService.getProducts();
+  }
+
+  onProductToCartAdded(product: ProductModel) {
+    this.cartService.addToCart(product);
   }
 }