+++ src/app/products/models/product.model.ts
@@ -1,11 +1,5 @@
 export class ProductModel {
-    // id: number;
-    // name: string;
-    // price: number;
-
+    
     constructor(public id: number, public name: string, public price: number) {
-        // this.id = id;
-        // this.name = name;
-        // this.price = price;
     }
 }